+++ NEWS.md
@@ -1,6 +1,6 @@
 # CHANGES IN VERSION 1.13.14
 
-* update translationalEfficiency with normalization by RNAseq level method. 
+* simplify the normBy and translationalEfficiency functions. 
 
 # CHANGES IN VERSION 1.13.13